pipeline/requirements.txt
pylint
pytest
pytest-cov
python-dotenv
psycopg2-binary

